% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 6 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3666|     0     |       0       |
#> |     Title     |  character  |            3666|     0     |       0       |
#> |      Beg      |    mdate    |            3666|    65     |     1.77      |
#> |    DocType    |  character  |            3666|     0     |       0       |
#> | AgreementType |  character  |            3666|     0     |       0       |
#> |   Signature   |    mdate    |            3666|    65     |     1.77      |
#> |     Force     |    mdate    |            3666|    867    |     23.65     |
#> |    Lineage    |  character  |            3666|     0     |       0       |
#> |   treatyID    |  character  |            3666|     0     |       0       |
#> |    ieadbID    |  character  |            3666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ECOLEX
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2174|     0     |       0       |
#> |     Title     |  character  |            2174|     0     |       0       |
#> |      Beg      |    mdate    |            2174|     0     |       0       |
#> |    DocType    |  character  |            2174|     0     |       0       |
#> |   GeogArea    |  character  |            2174|    90     |     4.14      |
#> |   Signature   |    mdate    |            2174|     0     |       0       |
#> |     Force     |    mdate    |            2174|   1066    |     49.03     |
#> |    Lineage    |  character  |            2174|     0     |       0       |
#> |   treatyID    |  character  |            2174|     0     |       0       |
#> |   ecolexID    |  character  |            2174|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $CIESIN
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             666|     0     |       0       |
#> |     Title     |  character  |             666|     0     |       0       |
#> |      Beg      |    mdate    |             666|     0     |       0       |
#> |   Signature   |    mdate    |             666|     0     |       0       |
#> |     Force     |    mdate    |             666|     0     |       0       |
#> |    Lineage    |  character  |             666|     0     |       0       |
#> |   treatyID    |  character  |             666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |      Beg      |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            3904|     0     |       0       |
#> |      Title      |  character  |            3904|     0     |       0       |
#> |       Beg       |    mdate    |            3904|    105    |     2.69      |
#> |       End       |    mdate    |            3904|   3235    |     82.86     |
#> |    Signature    |    mdate    |            3904|    105    |     2.69      |
#> |      Force      |    mdate    |            3904|   1745    |     44.7      |
#> |  AgreementType  |  character  |            3904|   1146    |     29.35     |
#> |     DocType     |  character  |            3904|   1108    |     28.38     |
#> |    GeogArea     |  character  |            3904|   1712    |     43.85     |
#> |     gengID      |  character  |            3904|     0     |       0       |
#> |      ieaID      |  character  |            3904|   3205    |     82.1      |
#> |    ecolexID     |  character  |            3904|   3421    |     87.63     |
#> |    treatyID     |  character  |            3904|     0     |       0       |
#> |     Parties     |  character  |            3904|   1687    |     43.21     |
#> |    verified     |  character  |            3904|   3855    |     98.74     |
#> |DocValidUntilDate|  character  |            3904|   3870    |     99.13     |
#> |       url       |  character  |            3904|   3132    |     80.23     |
#> |      Notes      |  character  |            3904|   3856    |     98.77     |
#> |    Download     |  character  |            3904|   2997    |     76.77     |
#> |    MEA_type     |  character  |            3904|   3205    |     82.1      |
#> |      Ambit      |  character  |            3904|   3479    |     89.11     |
#> |     Region      |  character  |            3904|   3655    |     93.62     |
#> | subject_ecolex  |  character  |            3904|   3421    |     87.63     |
#> |   subject_iea   |  character  |            3904|   3205    |     82.1      |
#> |    Keywords     |  character  |            3904|   3492    |     89.45     |
#> |     Lineage     |  character  |            3904|   3205    |     82.1      |
#> |    Sequence     |  character  |            3904|   3206    |     82.12     |
#> |    AdoptedIn    |  character  |            3904|   3473    |     88.96     |
#> |    Languages    |  character  |            3904|   3421    |     87.63     |
#> |   Appendices    |  character  |            3904|   3869    |     99.1      |
#> |   Depository    |  character  |            3904|   3515    |     90.04     |
#> |  DepositoryURL  |  character  |            3904|   3888    |     99.59     |
#> |    Published    |  character  |            3904|   3639    |     93.21     |
#> |    Abstract     |  character  |            3904|   3878    |     99.33     |
#> |    Website1     |  character  |            3904|   3520    |     90.16     |
#> |    Website2     |  character  |            3904|   3834    |     98.21     |
#> |   Secretariat   |  character  |            3904|   3206    |     82.12     |
#> | SecretariatURL  |  character  |            3904|   3206    |     82.12     |
#> |      UNEP       |  character  |            3904|   3703    |     94.85     |
#> |   Supersedes    |  character  |            3904|   3887    |     99.56     |
#> |   References    |  character  |            3904|   3868    |     99.08     |
#> |    EnabledBy    |  character  |            3904|   3796    |     97.23     |
#> |    AmendedBy    |  character  |            3904|   3812    |     97.64     |
#> |       Lit       |  character  |            3904|   3352    |     85.86     |
#> |      Data       |  character  |            3904|   3360    |     86.07     |
#> |      Coded      |  character  |            3904|   3600    |     92.21     |
#> ------------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> -----------------------------------------------------------------------------
#> |  Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -----------------------------------------------------------------------------
#> |     manyID     |  character  |            3815|     0     |       0       |
#> |    treatyID    |  character  |            3815|     0     |       0       |
#> |     Title      |  character  |            3815|     0     |       0       |
#> |      Beg       |    mdate    |            3815|    53     |     1.39      |
#> |      End       |    mdate    |            3815|   3142    |     82.36     |
#> |   Signature    |    mdate    |            3815|    53     |     1.39      |
#> |     Force      |    mdate    |            3815|   1671    |     43.8      |
#> |  action_area   |  character  |            3815|    426    |     11.17     |
#> |    linkage     |  character  |            3815|   2412    |     63.22     |
#> |   accessionC   |  character  |            3815|    622    |     16.3      |
#> |   accessionP   |  character  |            3815|   3553    |     93.13     |
#> |termination_type|  character  |            3815|   3412    |     89.44     |
#> |termination_date|  character  |            3815|   3722    |     97.56     |
#> -----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
