% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_agreements.R
\name{generate_}
\alias{generate_}
\alias{generate_agreements}
\alias{generate_memberships}
\title{Generate agreements}
\usage{
generate_agreements(n = 10)

generate_memberships(n = 10)
}
\arguments{
\item{n}{Integer number of fictional agreement names to generate.}
}
\value{
String vector of fictional agreement names and/or memberships.
}
\description{
Generate a list of fictional agreements and memberships.
}
\examples{
generate_agreements(12)
generate_memberships(12)
}
