% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_agreements.R
\name{generate_agreements}
\alias{generate_agreements}
\title{Generate a list of fictional agreement names}
\usage{
generate_agreements(n = 10)
}
\arguments{
\item{n}{Integer number of fictional agreement names to generate.}
}
\value{
String vector of fictional agreement names
}
\description{
Generate a list of fictional agreement names
}
\examples{
  generate_agreements(12)
}
