% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-package.R
\docType{package}
\name{manyenviron-package}
\alias{manyenviron}
\alias{manyenviron-package}
\title{manyenviron: Environmental agreements for the many universe}
\description{
Manyenviron provides an ensemble of data on international environmental agreements,
    and states membership in them for research use.
}
\author{
\strong{Maintainer}: James Hollway \email{james.hollway@graduateinstitute.ch} (\href{https://orcid.org/0000-0002-8361-9647}{ORCID}) (IHEID) [contributor]

Other contributors:
\itemize{
  \item Henrique Sposito (\href{https://orcid.org/0000-0003-3420-6085}{ORCID}) (IHEID) [contributor]
  \item Bernhard Bieri (\href{https://orcid.org/0000-0001-5943-9059}{ORCID}) (IHEID) [contributor]
  \item Esther Peev (\href{https://orcid.org/0000-0002-9678-2777}{ORCID}) (IHEID) [contributor]
}

}
\keyword{internal}
