% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-organizations.R
\docType{data}
\name{organizations}
\alias{organizations}
\title{organizations database documentation}
\format{
The organizations database is a list that contains the
following 2 datasets: MIA, YIO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{MIA: }{A dataset with 78 observations and the following
4 variables: Name, Beg, End, COW.}
\item{YIO: }{A dataset with 75115 observations and the following
4 variables: Title, Abbreviation, Beg, Country.}
}
}
\usage{
organizations
}
\description{
organizations database documentation
}
\keyword{datasets}
