% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qEnviron-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 1 datasets: REF.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{REF: }{A dataset with 2174 observations and the following
32 variables: Treaty1, Action, Treaty2, Treaty3, Treaty4, Treaty5,
Treaty6, Treaty7, Treaty8, Treaty9, Treaty10, Treaty11, Treaty12,
Treaty13, Treaty14, Treaty15, Treaty16, Treaty17, Treaty18, Treaty19,
Treaty20, Treaty21, Treaty22, Treaty23, Treaty24, Treaty25, Treaty26,
Treaty27, Treaty28, Treaty29, Treaty30, Treaty31.}
}
}
\usage{
references
}
\description{
references database documentation
}
\keyword{datasets}
