% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-regimes.R
\docType{data}
\name{regimes}
\alias{regimes}
\title{regimes database documentation}
\format{
The regimes database is a list that contains the
following 1 datasets: IRD.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IRD: }{A dataset with 92 observations and the following
7 variables: Regime, RegimeComponent, Beg, Wat, Wat2, End, RegimeElement.}
}
}
\usage{
regimes
}
\description{
regimes database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
