% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 6 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IEADB: }{A dataset with 3666 observations and the following
10 variables: manyID, Title, Beg, DocType, AgreementType, Signature, Force,
Lineage, treatyID, ieadbID.}
\item{ECOLEX: }{A dataset with 2174 observations and the following
10 variables: manyID, Title, Beg, DocType, GeogArea, Signature, Force,
Lineage, treatyID, ecolexID.}
\item{CIESIN: }{A dataset with 666 observations and the following
7 variables: manyID, Title, Beg, Signature, Force, Lineage, treatyID.}
\item{HEIDI: }{A dataset with 2280 observations and the following
7 variables: manyID, Title, Beg, Signature, Lineage, treatyID, heidiID.}
\item{HUGGO: }{A dataset with 3953 observations and the following
54 variables: manyID, Title, Beg, End, Signature, Force, AgreementType,
DocType, GeogArea, gengID, ieaID, ecolexID, treatyID, Parties, verified,
DocValidUntilDate, url, Notes, Download, MEA_type, Ambit, Region,
subject_ecolex, subject_iea, Keywords, Lineage, Sequence, AdoptedIn,
Languages, Appendices, Depository, DepositoryURL, Published, Abstract,
Website1, Website2, Secretariat, SecretariatURL, UNEP, Supersedes,
References, EnabledBy, AmendedBy, Lit, Data, Coded, TreatyText, Source,
MainText, AppendixText, AnnexText, Checked_HUGGO, Confirmed_HUGGO, Changes.}
\item{AIGGO: }{A dataset with 3815 observations and the following
13 variables: manyID, treatyID, Title, Beg, End, Signature, Force,
action_area, linkage, accessionC, accessionP, termination_type,
termination_date.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3666|     0     |       0       |
#> |     Title     |  character  |            3666|     0     |       0       |
#> |      Beg      |    mdate    |            3666|    65     |     1.77      |
#> |    DocType    |  character  |            3666|     0     |       0       |
#> | AgreementType |  character  |            3666|     0     |       0       |
#> |   Signature   |    mdate    |            3666|    65     |     1.77      |
#> |     Force     |    mdate    |            3666|    867    |     23.65     |
#> |    Lineage    |  character  |            3666|     0     |       0       |
#> |   treatyID    |  character  |            3666|     0     |       0       |
#> |    ieadbID    |  character  |            3666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ECOLEX
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2174|     0     |       0       |
#> |     Title     |  character  |            2174|     0     |       0       |
#> |      Beg      |    mdate    |            2174|     0     |       0       |
#> |    DocType    |  character  |            2174|     0     |       0       |
#> |   GeogArea    |  character  |            2174|    90     |     4.14      |
#> |   Signature   |    mdate    |            2174|     0     |       0       |
#> |     Force     |    mdate    |            2174|   1066    |     49.03     |
#> |    Lineage    |  character  |            2174|     0     |       0       |
#> |   treatyID    |  character  |            2174|     0     |       0       |
#> |   ecolexID    |  character  |            2174|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $CIESIN
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             666|     0     |       0       |
#> |     Title     |  character  |             666|     0     |       0       |
#> |      Beg      |    mdate    |             666|     0     |       0       |
#> |   Signature   |    mdate    |             666|     0     |       0       |
#> |     Force     |    mdate    |             666|     0     |       0       |
#> |    Lineage    |  character  |             666|     0     |       0       |
#> |   treatyID    |  character  |             666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |      Beg      |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            3953|     0     |       0       |
#> |      Title      |  character  |            3953|     0     |       0       |
#> |       Beg       |    mdate    |            3953|    105    |     2.66      |
#> |       End       |    mdate    |            3953|   3235    |     81.84     |
#> |    Signature    |    mdate    |            3953|    105    |     2.66      |
#> |      Force      |    mdate    |            3953|   1713    |     43.33     |
#> |  AgreementType  |  character  |            3953|   1138    |     28.79     |
#> |     DocType     |  character  |            3953|   1100    |     27.83     |
#> |    GeogArea     |  character  |            3953|   1716    |     43.41     |
#> |     gengID      |  character  |            3953|     0     |       0       |
#> |      ieaID      |  character  |            3953|   3227    |     81.63     |
#> |    ecolexID     |  character  |            3953|   3461    |     87.55     |
#> |    treatyID     |  character  |            3953|     0     |       0       |
#> |     Parties     |  character  |            3953|   1695    |     42.88     |
#> |    verified     |  character  |            3953|   3902    |     98.71     |
#> |DocValidUntilDate|  character  |            3953|   3918    |     99.11     |
#> |       url       |  character  |            3953|   3132    |     79.23     |
#> |      Notes      |  character  |            3953|   3903    |     98.74     |
#> |    Download     |  character  |            3953|   3017    |     76.32     |
#> |    MEA_type     |  character  |            3953|   3227    |     81.63     |
#> |      Ambit      |  character  |            3953|   3521    |     89.07     |
#> |     Region      |  character  |            3953|   3698    |     93.55     |
#> | subject_ecolex  |  character  |            3953|   3461    |     87.55     |
#> |   subject_iea   |  character  |            3953|   3227    |     81.63     |
#> |    Keywords     |  character  |            3953|   3534    |     89.4      |
#> |     Lineage     |  character  |            3953|   3227    |     81.63     |
#> |    Sequence     |  character  |            3953|   3228    |     81.66     |
#> |    AdoptedIn    |  character  |            3953|   3513    |     88.87     |
#> |    Languages    |  character  |            3953|   3461    |     87.55     |
#> |   Appendices    |  character  |            3953|   3917    |     99.09     |
#> |   Depository    |  character  |            3953|   3557    |     89.98     |
#> |  DepositoryURL  |  character  |            3953|   3936    |     99.57     |
#> |    Published    |  character  |            3953|   3682    |     93.14     |
#> |    Abstract     |  character  |            3953|   3926    |     99.32     |
#> |    Website1     |  character  |            3953|   3562    |     90.11     |
#> |    Website2     |  character  |            3953|   3883    |     98.23     |
#> |   Secretariat   |  character  |            3953|   3228    |     81.66     |
#> | SecretariatURL  |  character  |            3953|   3228    |     81.66     |
#> |      UNEP       |  character  |            3953|   3745    |     94.74     |
#> |   Supersedes    |  character  |            3953|   3936    |     99.57     |
#> |   References    |  character  |            3953|   3917    |     99.09     |
#> |    EnabledBy    |  character  |            3953|   3842    |     97.19     |
#> |    AmendedBy    |  character  |            3953|   3859    |     97.62     |
#> |       Lit       |  character  |            3953|   3380    |     85.5      |
#> |      Data       |  character  |            3953|   3390    |     85.76     |
#> |      Coded      |  character  |            3953|   3645    |     92.21     |
#> |   TreatyText    |  character  |            3953|   2858    |     72.3      |
#> |     Source      |  character  |            3953|     0     |       0       |
#> |    MainText     |  character  |            3953|   2839    |     71.82     |
#> |  AppendixText   |  character  |            3953|   3880    |     98.15     |
#> |    AnnexText    |  character  |            3953|   3565    |     90.18     |
#> |  Checked_HUGGO  |   numeric   |            3953|   3932    |     99.47     |
#> | Confirmed_HUGGO |  character  |            3953|   3932    |     99.47     |
#> |     Changes     |  character  |            3953|   3932    |     99.47     |
#> ------------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> -----------------------------------------------------------------------------
#> |  Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -----------------------------------------------------------------------------
#> |     manyID     |  character  |            3815|     0     |       0       |
#> |    treatyID    |  character  |            3815|     0     |       0       |
#> |     Title      |  character  |            3815|     0     |       0       |
#> |      Beg       |    mdate    |            3815|    53     |     1.39      |
#> |      End       |    mdate    |            3815|   3142    |     82.36     |
#> |   Signature    |    mdate    |            3815|    53     |     1.39      |
#> |     Force      |    mdate    |            3815|   1671    |     43.8      |
#> |  action_area   |  character  |            3815|    426    |     11.17     |
#> |    linkage     |  character  |            3815|   2412    |     63.22     |
#> |   accessionC   |  character  |            3815|    622    |     16.3      |
#> |   accessionP   |  character  |            3815|   3553    |     93.13     |
#> |termination_type|  character  |            3815|   3412    |     89.44     |
#> |termination_date|  character  |            3815|   3722    |     97.56     |
#> -----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
