% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 5 datasets: IEADB, GNEVAR, ECOLEX, CIESIN, HEIDI.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IEADB: }{A dataset with 3666 observations and the following
10 variables: manyID, Title, Beg, L, D, Signature, Force, Lineage,
treatyID, ieadbID.}
\item{GNEVAR: }{A dataset with 3428 observations and the following
12 variables: manyID, Title, Beg, End, L, D, J, Signature, Force,
Lineage, treatyID, gnevarID.}
\item{ECOLEX: }{A dataset with 2174 observations and the following
10 variables: manyID, Title, Beg, L, J, Signature, Force, Lineage,
treatyID, ecolexID.}
\item{CIESIN: }{A dataset with 666 observations and the following
7 variables: manyID, Title, Beg, Signature, Force, Lineage, treatyID.}
\item{HEIDI: }{A dataset with 2280 observations and the following
7 variables: manyID, Title, Beg, Signature, Lineage, treatyID, heidiID.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
