% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 6 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $IEADB
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            3666|     0     |       0       |
## |     Title     |  character  |            3666|     0     |       0       |
## |      Beg      |    mdate    |            3666|    65     |     1.77      |
## |    DocType    |  character  |            3666|     0     |       0       |
## | AgreementType |  character  |            3666|     0     |       0       |
## |   Signature   |    mdate    |            3666|    65     |     1.77      |
## |     Force     |    mdate    |            3666|    867    |     23.65     |
## |    Lineage    |  character  |            3666|     0     |       0       |
## |   treatyID    |  character  |            3666|     0     |       0       |
## |    ieadbID    |  character  |            3666|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $ECOLEX
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            2174|     0     |       0       |
## |     Title     |  character  |            2174|     0     |       0       |
## |      Beg      |    mdate    |            2174|     0     |       0       |
## |    DocType    |  character  |            2174|     0     |       0       |
## |   GeogArea    |  character  |            2174|    90     |     4.14      |
## |   Signature   |    mdate    |            2174|     0     |       0       |
## |     Force     |    mdate    |            2174|   1066    |     49.03     |
## |    Lineage    |  character  |            2174|     0     |       0       |
## |   treatyID    |  character  |            2174|     0     |       0       |
## |   ecolexID    |  character  |            2174|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $CIESIN
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |             666|     0     |       0       |
## |     Title     |  character  |             666|     0     |       0       |
## |      Beg      |    mdate    |             666|     0     |       0       |
## |   Signature   |    mdate    |             666|     0     |       0       |
## |     Force     |    mdate    |             666|     0     |       0       |
## |    Lineage    |  character  |             666|     0     |       0       |
## |   treatyID    |  character  |             666|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $HEIDI
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            2280|     0     |       0       |
## |     Title     |  character  |            2280|     0     |       0       |
## |      Beg      |    mdate    |            2280|    27     |     1.18      |
## |   Signature   |    mdate    |            2280|    27     |     1.18      |
## |    Lineage    |  character  |            2280|     0     |       0       |
## |   treatyID    |  character  |            2280|     0     |       0       |
## |    heidiID    |   numeric   |            2280|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $HUGGO
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            8213|    825    |     10.05     |
## |     Title     |  character  |            8213|     0     |       0       |
## |      Beg      |    mdate    |            8213|    457    |     5.56      |
## |      End      |    mdate    |            8213|   7526    |     91.64     |
## |    DocType    |  character  |            8213|   5353    |     65.18     |
## | AgreementType |  character  |            8213|   5391    |     65.64     |
## |   GeogArea    |  character  |            8213|   6022    |     73.32     |
## |   Signature   |    mdate    |            8213|   4710    |     57.35     |
## |     Force     |    mdate    |            8213|   6306    |     76.78     |
## |   treatyID    |  character  |            8213|   4653    |     56.65     |
## |   gnevarID    |  character  |            8213|   5357    |     65.23     |
## |  TreatyText   |    list     |            8213|   4120    |     50.16     |
## |    Source     |    list     |            8213|   5037    |     61.33     |
## ----------------------------------------------------------------------------
## 
## 
## $AIGGO
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            8213|    825    |     10.05     |
## |     Title     |  character  |            8213|     0     |       0       |
## |      Beg      |    mdate    |            8213|    457    |     5.56      |
## |      End      |    mdate    |            8213|   7526    |     91.64     |
## |    DocType    |  character  |            8213|   5353    |     65.18     |
## | AgreementType |  character  |            8213|   5391    |     65.64     |
## |   GeogArea    |  character  |            8213|   6022    |     73.32     |
## |   Signature   |    mdate    |            8213|   4710    |     57.35     |
## |     Force     |    mdate    |            8213|   6306    |     76.78     |
## |    Lineage    |  character  |            8213|     0     |       0       |
## |  accessionC   |  character  |            8213|   1441    |     17.55     |
## |  accessionP   |  character  |            8213|   7585    |     92.35     |
## |   treatyID    |  character  |            8213|   4653    |     56.65     |
## |   gnevarID    |  character  |            8213|   5357    |     65.23     |
## |  TreatyText   |    list     |            8213|   4120    |     50.16     |
## |    Source     |    list     |            8213|   5037    |     61.33     |
## ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
