% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 6 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IEADB: }{A dataset with 3666 observations and the following
10 variables: manyID, treatyID, Title, Beg, DocType, AgreementType,
Signature, Force, Lineage, ieadbID.}
\item{ECOLEX: }{A dataset with 2174 observations and the following
10 variables: manyID, treatyID, Title, Beg, DocType, GeogArea, Signature,
Force, Lineage, ecolexID.}
\item{CIESIN: }{A dataset with 662 observations and the following
7 variables: manyID, treatyID, Title, Beg, Signature, Force, Lineage.}
\item{HEIDI: }{A dataset with 2280 observations and the following
7 variables: manyID, treatyID, Title, Beg, Signature, Lineage, heidiID.}
\item{HUGGO: }{A dataset with 4036 observations and the following
48 variables: manyID, treatyID, Title, Beg, End, Signature, Force, url,
AgreementType, DocType, GeogArea, gengID, ieaID, ecolexID, Parties, verified,
DocValidUntilDate, Notes, Download, MEA_type, Ambit, Region, subject_ecolex,
subject_iea, Keywords, Lineage, Sequence, AdoptedIn, Languages, Appendices,
Depository, DepositoryURL, Published, Website1, Website2, Secretariat,
SecretariatURL, UNEP, Supersedes, References, EnabledBy, AmendedBy, Lit,
Data, Coded, Abstract, TreatyText, Language.}
\item{AIGGO: }{A dataset with 3778 observations and the following
13 variables: manyID, treatyID, Title, Beg, End, Signature, Force,
action_area, linkage, accessionC, accessionP, termination_type,
termination_date.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3666|     0     |       0       |
#> |   treatyID    |  character  |            3666|     0     |       0       |
#> |     Title     |  character  |            3666|     0     |       0       |
#> |      Beg      |    mdate    |            3666|     0     |       0       |
#> |    DocType    |  character  |            3666|     0     |       0       |
#> | AgreementType |  character  |            3666|     0     |       0       |
#> |   Signature   |    mdate    |            3666|    65     |     1.77      |
#> |     Force     |    mdate    |            3666|    867    |     23.65     |
#> |    Lineage    |  character  |            3666|     0     |       0       |
#> |    ieadbID    |  character  |            3666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ECOLEX
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2174|     0     |       0       |
#> |   treatyID    |  character  |            2174|     0     |       0       |
#> |     Title     |  character  |            2174|     0     |       0       |
#> |      Beg      |    mdate    |            2174|     0     |       0       |
#> |    DocType    |  character  |            2174|     0     |       0       |
#> |   GeogArea    |  character  |            2174|    90     |     4.14      |
#> |   Signature   |    mdate    |            2174|     0     |       0       |
#> |     Force     |    mdate    |            2174|   1066    |     49.03     |
#> |    Lineage    |  character  |            2174|     0     |       0       |
#> |   ecolexID    |  character  |            2174|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $CIESIN
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             662|     0     |       0       |
#> |   treatyID    |  character  |             662|     0     |       0       |
#> |     Title     |  character  |             662|     0     |       0       |
#> |      Beg      |    mdate    |             662|     0     |       0       |
#> |   Signature   |    mdate    |             662|     0     |       0       |
#> |     Force     |    mdate    |             662|     0     |       0       |
#> |    Lineage    |  character  |             662|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |      Beg      |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            3980|     0     |       0       |
#> |    treatyID     |  character  |            3980|     0     |       0       |
#> |      Title      |  character  |            3980|     0     |       0       |
#> |       Beg       |    mdate    |            3980|    57     |     1.43      |
#> |       End       |    mdate    |            3980|   3291    |     82.69     |
#> |    Signature    |    mdate    |            3980|    57     |     1.43      |
#> |      Force      |    mdate    |            3980|   1616    |     40.6      |
#> |       url       |  character  |            3980|   2688    |     67.54     |
#> |  AgreementType  |  character  |            3980|   1279    |     32.14     |
#> |     DocType     |  character  |            3980|   1243    |     31.23     |
#> |    GeogArea     |  character  |            3980|   1842    |     46.28     |
#> |     gengID      |  character  |            3980|    195    |      4.9      |
#> |      ieaID      |  character  |            3980|   3284    |     82.51     |
#> |    ecolexID     |  character  |            3980|   3503    |     88.02     |
#> |     Parties     |  character  |            3980|   1709    |     42.94     |
#> |    verified     |  character  |            3980|   3932    |     98.79     |
#> |DocValidUntilDate|  character  |            3980|   3946    |     99.15     |
#> |      Notes      |  character  |            3980|   3933    |     98.82     |
#> |    Download     |  character  |            3980|   3081    |     77.41     |
#> |    MEA_type     |  character  |            3980|   3284    |     82.51     |
#> |      Ambit      |  character  |            3980|   3561    |     89.47     |
#> |     Region      |  character  |            3980|   3733    |     93.79     |
#> | subject_ecolex  |  character  |            3980|   3503    |     88.02     |
#> |   subject_iea   |  character  |            3980|   3284    |     82.51     |
#> |    Keywords     |  character  |            3980|   3573    |     89.77     |
#> |     Lineage     |  character  |            3980|   3284    |     82.51     |
#> |    Sequence     |  character  |            3980|   3285    |     82.54     |
#> |    AdoptedIn    |  character  |            3980|   3553    |     89.27     |
#> |    Languages    |  character  |            3980|   3503    |     88.02     |
#> |   Appendices    |  character  |            3980|   3945    |     99.12     |
#> |   Depository    |  character  |            3980|   3593    |     90.28     |
#> |  DepositoryURL  |  character  |            3980|   3964    |     99.6      |
#> |    Published    |  character  |            3980|   3719    |     93.44     |
#> |    Website1     |  character  |            3980|   3598    |     90.4      |
#> |    Website2     |  character  |            3980|   3956    |     99.4      |
#> |   Secretariat   |  character  |            3980|   3285    |     82.54     |
#> | SecretariatURL  |  character  |            3980|   3285    |     82.54     |
#> |      UNEP       |  character  |            3980|   3778    |     94.92     |
#> |   Supersedes    |  character  |            3980|   3963    |     99.57     |
#> |   References    |  character  |            3980|   3945    |     99.12     |
#> |    EnabledBy    |  character  |            3980|   3873    |     97.31     |
#> |    AmendedBy    |  character  |            3980|   3889    |     97.71     |
#> |       Lit       |  character  |            3980|   3435    |     86.31     |
#> |      Data       |  character  |            3980|   3442    |     86.48     |
#> |      Coded      |  character  |            3980|   3681    |     92.49     |
#> |    Abstract     |  character  |            3980|   3954    |     99.35     |
#> |   TreatyText    |   numeric   |            3980|   2632    |     66.13     |
#> |    Language     |  character  |            3980|   2632    |     66.13     |
#> |Orig_noneng_title|  character  |            3980|   3924    |     98.59     |
#> |      match      |  character  |            3980|   3924    |     98.59     |
#> ------------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> -----------------------------------------------------------------------------
#> |  Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -----------------------------------------------------------------------------
#> |     manyID     |  character  |            3778|     0     |       0       |
#> |    treatyID    |  character  |            3778|     0     |       0       |
#> |     Title      |  character  |            3778|     0     |       0       |
#> |      Beg       |    mdate    |            3778|    47     |     1.24      |
#> |      End       |    mdate    |            3778|   3112    |     82.37     |
#> |   Signature    |    mdate    |            3778|    47     |     1.24      |
#> |     Force      |    mdate    |            3778|   1646    |     43.57     |
#> |  action_area   |  character  |            3778|    415    |     10.98     |
#> |    linkage     |  character  |            3778|   2401    |     63.55     |
#> |   accessionC   |  character  |            3778|    610    |     16.15     |
#> |   accessionP   |  character  |            3778|   3516    |     93.07     |
#> |termination_type|  character  |            3778|   3375    |     89.33     |
#> |termination_date|  character  |            3778|   3685    |     97.54     |
#> -----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
