% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 6 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IEADB: }{A dataset with 3666 observations and the following
10 variables: manyID, Title, Beg, DocType, AgreementType, Signature, Force, Lineage, treatyID, ieadbID.}
\item{ECOLEX: }{A dataset with 2174 observations and the following
10 variables: manyID, Title, Beg, DocType, GeogArea, Signature, Force, Lineage, treatyID, ecolexID.}
\item{CIESIN: }{A dataset with 666 observations and the following
7 variables: manyID, Title, Beg, Signature, Force, Lineage, treatyID.}
\item{HEIDI: }{A dataset with 2280 observations and the following
7 variables: manyID, Title, Beg, Signature, Lineage, treatyID, heidiID.}
\item{HUGGO: }{A dataset with 4084 observations and the following
48 variables: manyID, Title, Beg, End, Signature, Force, url, AgreementType, DocType, GeogArea, gengID, ieaID, ecolexID, treatyID, Parties, verified, DocValidUntilDate, Notes, Download, MEA_type, Ambit, Region, subject_ecolex, subject_iea, Keywords, Lineage, Sequence, AdoptedIn, Languages, Appendices, Depository, DepositoryURL, Published, Abstract, Website1, Website2, Secretariat, SecretariatURL, UNEP, Supersedes, References, EnabledBy, AmendedBy, Lit, Data, Coded, TreatyText, Language.}
\item{AIGGO: }{A dataset with 3815 observations and the following
13 variables: manyID, treatyID, Title, Beg, End, Signature, Force, action_area, linkage, accessionC, accessionP, termination_type, termination_date.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3666|     0     |       0       |
#> |     Title     |  character  |            3666|     0     |       0       |
#> |      Beg      |    mdate    |            3666|    65     |     1.77      |
#> |    DocType    |  character  |            3666|     0     |       0       |
#> | AgreementType |  character  |            3666|     0     |       0       |
#> |   Signature   |    mdate    |            3666|    65     |     1.77      |
#> |     Force     |    mdate    |            3666|    867    |     23.65     |
#> |    Lineage    |  character  |            3666|     0     |       0       |
#> |   treatyID    |  character  |            3666|     0     |       0       |
#> |    ieadbID    |  character  |            3666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ECOLEX
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2174|     0     |       0       |
#> |     Title     |  character  |            2174|     0     |       0       |
#> |      Beg      |    mdate    |            2174|     0     |       0       |
#> |    DocType    |  character  |            2174|     0     |       0       |
#> |   GeogArea    |  character  |            2174|    90     |     4.14      |
#> |   Signature   |    mdate    |            2174|     0     |       0       |
#> |     Force     |    mdate    |            2174|   1066    |     49.03     |
#> |    Lineage    |  character  |            2174|     0     |       0       |
#> |   treatyID    |  character  |            2174|     0     |       0       |
#> |   ecolexID    |  character  |            2174|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $CIESIN
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             666|     0     |       0       |
#> |     Title     |  character  |             666|     0     |       0       |
#> |      Beg      |    mdate    |             666|     0     |       0       |
#> |   Signature   |    mdate    |             666|     0     |       0       |
#> |     Force     |    mdate    |             666|     0     |       0       |
#> |    Lineage    |  character  |             666|     0     |       0       |
#> |   treatyID    |  character  |             666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |      Beg      |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            4084|     0     |       0       |
#> |      Title      |  character  |            4084|     0     |       0       |
#> |       Beg       |    mdate    |            4084|    105    |     2.57      |
#> |       End       |    mdate    |            4084|   3393    |     83.08     |
#> |    Signature    |    mdate    |            4084|    105    |     2.57      |
#> |      Force      |    mdate    |            4084|   1695    |     41.5      |
#> |       url       |  character  |            4084|   2790    |     68.32     |
#> |  AgreementType  |  character  |            4084|   1335    |     32.69     |
#> |     DocType     |  character  |            4084|   1299    |     31.81     |
#> |    GeogArea     |  character  |            4084|   1899    |     46.5      |
#> |     gengID      |  character  |            4084|    195    |     4.77      |
#> |      ieaID      |  character  |            4084|   3388    |     82.96     |
#> |    ecolexID     |  character  |            4084|   3603    |     88.22     |
#> |    treatyID     |  character  |            4084|    195    |     4.77      |
#> |     Parties     |  character  |            4084|   1761    |     43.12     |
#> |    verified     |  character  |            4084|   4036    |     98.82     |
#> |DocValidUntilDate|  character  |            4084|   4050    |     99.17     |
#> |      Notes      |  character  |            4084|   4037    |     98.85     |
#> |    Download     |  character  |            4084|   3181    |     77.89     |
#> |    MEA_type     |  character  |            4084|   3388    |     82.96     |
#> |      Ambit      |  character  |            4084|   3661    |     89.64     |
#> |     Region      |  character  |            4084|   3836    |     93.93     |
#> | subject_ecolex  |  character  |            4084|   3603    |     88.22     |
#> |   subject_iea   |  character  |            4084|   3388    |     82.96     |
#> |    Keywords     |  character  |            4084|   3674    |     89.96     |
#> |     Lineage     |  character  |            4084|   3388    |     82.96     |
#> |    Sequence     |  character  |            4084|   3389    |     82.98     |
#> |    AdoptedIn    |  character  |            4084|   3653    |     89.45     |
#> |    Languages    |  character  |            4084|   3603    |     88.22     |
#> |   Appendices    |  character  |            4084|   4049    |     99.14     |
#> |   Depository    |  character  |            4084|   3695    |     90.48     |
#> |  DepositoryURL  |  character  |            4084|   4068    |     99.61     |
#> |    Published    |  character  |            4084|   3820    |     93.54     |
#> |    Abstract     |  character  |            4084|   4058    |     99.36     |
#> |    Website1     |  character  |            4084|   3701    |     90.62     |
#> |    Website2     |  character  |            4084|   4060    |     99.41     |
#> |   Secretariat   |  character  |            4084|   3389    |     82.98     |
#> | SecretariatURL  |  character  |            4084|   3389    |     82.98     |
#> |      UNEP       |  character  |            4084|   3882    |     95.05     |
#> |   Supersedes    |  character  |            4084|   4067    |     99.58     |
#> |   References    |  character  |            4084|   4049    |     99.14     |
#> |    EnabledBy    |  character  |            4084|   3976    |     97.36     |
#> |    AmendedBy    |  character  |            4084|   3993    |     97.77     |
#> |       Lit       |  character  |            4084|   3535    |     86.56     |
#> |      Data       |  character  |            4084|   3542    |     86.73     |
#> |      Coded      |  character  |            4084|   3781    |     92.58     |
#> |   TreatyText    |   numeric   |            4084|   2735    |     66.97     |
#> |    Language     |  character  |            4084|   2742    |     67.14     |
#> ------------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> -----------------------------------------------------------------------------
#> |  Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -----------------------------------------------------------------------------
#> |     manyID     |  character  |            3815|     0     |       0       |
#> |    treatyID    |  character  |            3815|     0     |       0       |
#> |     Title      |  character  |            3815|     0     |       0       |
#> |      Beg       |    mdate    |            3815|    53     |     1.39      |
#> |      End       |    mdate    |            3815|   3142    |     82.36     |
#> |   Signature    |    mdate    |            3815|    53     |     1.39      |
#> |     Force      |    mdate    |            3815|   1671    |     43.8      |
#> |  action_area   |  character  |            3815|    426    |     11.17     |
#> |    linkage     |  character  |            3815|   2412    |     63.22     |
#> |   accessionC   |  character  |            3815|    622    |     16.3      |
#> |   accessionP   |  character  |            3815|   3553    |     93.13     |
#> |termination_type|  character  |            3815|   3412    |     89.44     |
#> |termination_date|  character  |            3815|   3722    |     97.56     |
#> -----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
