% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qData-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the following  datasets: IEADB, GNEVAR, ECOLEX. For more info, please use the report_data() function.
\describe{
\item{IEADB: A dataset with 3666 observations and the following 3 variables: IEADB_ID, Title, Beg.}{Describe dataset here}
\item{GNEVAR: A dataset with 2733 observations and the following 4 variables: GENG_ID, Title, Beg, End.}{Describe dataset here}
\item{ECOLEX: A dataset with 2174 observations and the following 3 variables: ECOLEX_ID, Title, Beg.}{Describe dataset here}
}
}
\usage{
agreements
}
\description{
Write the database citation here.
}
\keyword{datasets}
