% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements datacube documentation}
\format{
The agreements datacube is a list that contains the
following 7 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO, IRD.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IEADB: }{A dataset with 3667 observations and the following
10 variables: manyID, Title, Begin, DocType, AgreementType, Signature, Force,
Lineage, treatyID, ieadbID.}
\item{ECOLEX: }{A dataset with 2174 observations and the following
10 variables: manyID, Title, Begin, DocType, GeogArea, Signature, Force,
Lineage, treatyID, ecolexID.}
\item{CIESIN: }{A dataset with 666 observations and the following
7 variables: manyID, Title, Begin, Signature, Force, Lineage, treatyID.}
\item{HEIDI: }{A dataset with 2280 observations and the following
7 variables: manyID, Title, Begin, Signature, Lineage, treatyID, heidiID.}
\item{HUGGO: }{A dataset with 3980 observations and the following
50 variables: manyID, treatyID, Title, Begin, End, Signature, Force, url,
AgreementType, DocType, GeogArea, gengID, ieaID, ecolexID, Parties, verified,
DocValidUntilDate, Notes, Download, MEA_type, Ambit, Region, subject_ecolex,
subject_iea, Keywords, Lineage, Sequence, AdoptedIn, Languages, Appendices,
Depository, DepositoryURL, Published, Website1, Website2, Secretariat,
SecretariatURL, UNEP, Supersedes, References, EnabledBy, AmendedBy, Lit,
Data, Coded, Abstract, TreatyText, Language, Orig_noneng_title, match.}
\item{AIGGO: }{A dataset with 3923 observations and the following
13 variables: manyID, treatyID, Title, Begin, End, Signature, Force,
action_area, linkage, accessionC, accessionP, termination_type,
termination_date.}
\item{IRD: }{A dataset with 92 observations and the following
9 variables: manyID, treatyID, Title, Begin, RegimeComponent, Wat, Wat2, End,
RegimeElement.}
}
}
\source{
\itemize{
\item{IEADB: }{
R. B. Mitchell et al. "What we know (and could know) about international environmental agreements". In: \emph{Global Environmental Politics} 20.1 (2020), pp. 103-121.
}
\item{ECOLEX: }{
J. Sommer. “Global governance in forestry: a cross-national analysis”. In: \emph{International Journal ofSustainable Development & World Ecology} 27.6 (2020), pp. 481-495.
}
\item{CIESIN: }{
Centre for International Earth Science Information Network. In: \emph{Socioeconomic Data and Applications Center (SEDAC) Collection of Treaty Texts} (2002).
}
\item{HEIDI: }{
J-F. Morin, and C. Blouin. "How environmental treaties contribute to global health governance". In: \emph{Globalization and health} 15.1 (2019), pp. 1-8.
}
\item{HUGGO: }{
J. Hollway. "Environmental agreements for manydata". (2021).}
\item{AIGGO: }{
J. Hollway. "Environmental agreements for manydata". (2021).}
\item{IRD: }{
O. R. Young and M. Zürn. “The international regimes database: Designing and using a sophisticated tool for institutional analysis”. In: \emph{Global Environmental Politics} 6.3 (2006), pp. 121-143.
}
}
}
\usage{
agreements
}
\description{
agreements datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3667|     0     |       0       |
#> |   treatyID    |  character  |            3667|     0     |       0       |
#> |     Title     |  character  |            3667|     0     |       0       |
#> |     Begin     |    mdate    |            3667|     0     |       0       |
#> |    DocType    |  character  |            3667|     0     |       0       |
#> | AgreementType |  character  |            3667|     0     |       0       |
#> |   Signature   |    mdate    |            3667|    65     |     1.77      |
#> |     Force     |    mdate    |            3667|    868    |     23.67     |
#> |    Lineage    |  character  |            3667|     0     |       0       |
#> |    ieadbID    |  character  |            3667|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ECOLEX
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2174|     0     |       0       |
#> |   treatyID    |  character  |            2174|     0     |       0       |
#> |     Title     |  character  |            2174|     0     |       0       |
#> |     Begin     |    mdate    |            2174|     0     |       0       |
#> |    DocType    |  character  |            2174|     0     |       0       |
#> |   GeogArea    |  character  |            2174|    90     |     4.14      |
#> |   Signature   |    mdate    |            2174|     0     |       0       |
#> |     Force     |    mdate    |            2174|   1066    |     49.03     |
#> |    Lineage    |  character  |            2174|     0     |       0       |
#> |   ecolexID    |  character  |            2174|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $CIESIN
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             662|     0     |       0       |
#> |   treatyID    |  character  |             662|     0     |       0       |
#> |     Title     |  character  |             662|     0     |       0       |
#> |     Begin     |    mdate    |             662|     0     |       0       |
#> |   Signature   |    mdate    |             662|     0     |       0       |
#> |     Force     |    mdate    |             662|     0     |       0       |
#> |    Lineage    |  character  |             662|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |     Begin     |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            3980|     0     |       0       |
#> |    treatyID     |  character  |            3980|     0     |       0       |
#> |      Title      |  character  |            3980|     0     |       0       |
#> |      Begin      |    mdate    |            3980|    57     |     1.43      |
#> |       End       |    mdate    |            3980|   3047    |     76.56     |
#> |    Signature    |    mdate    |            3980|    57     |     1.43      |
#> |      Force      |    mdate    |            3980|   1582    |     39.75     |
#> |       url       |  character  |            3980|   2688    |     67.54     |
#> |  AgreementType  |  character  |            3980|   1279    |     32.14     |
#> |     DocType     |  character  |            3980|   1243    |     31.23     |
#> |    GeogArea     |  character  |            3980|   1842    |     46.28     |
#> |     gengID      |  character  |            3980|    195    |      4.9      |
#> |      ieaID      |  character  |            3980|   3284    |     82.51     |
#> |    ecolexID     |  character  |            3980|   3503    |     88.02     |
#> |     Parties     |  character  |            3980|   1709    |     42.94     |
#> |    verified     |  character  |            3980|   3932    |     98.79     |
#> |DocValidUntilDate|  character  |            3980|   3946    |     99.15     |
#> |      Notes      |  character  |            3980|   3933    |     98.82     |
#> |    Download     |  character  |            3980|   3081    |     77.41     |
#> |    MEA_type     |  character  |            3980|   3284    |     82.51     |
#> |      Ambit      |  character  |            3980|   3561    |     89.47     |
#> |     Region      |  character  |            3980|   3733    |     93.79     |
#> | subject_ecolex  |  character  |            3980|   3503    |     88.02     |
#> |   subject_iea   |  character  |            3980|   3284    |     82.51     |
#> |    Keywords     |  character  |            3980|   3573    |     89.77     |
#> |     Lineage     |  character  |            3980|   3284    |     82.51     |
#> |    Sequence     |  character  |            3980|   3285    |     82.54     |
#> |    AdoptedIn    |  character  |            3980|   3553    |     89.27     |
#> |    Languages    |  character  |            3980|   3503    |     88.02     |
#> |   Appendices    |  character  |            3980|   3945    |     99.12     |
#> |   Depository    |  character  |            3980|   3593    |     90.28     |
#> |  DepositoryURL  |  character  |            3980|   3964    |     99.6      |
#> |    Published    |  character  |            3980|   3719    |     93.44     |
#> |    Website1     |  character  |            3980|   3598    |     90.4      |
#> |    Website2     |  character  |            3980|   3956    |     99.4      |
#> |   Secretariat   |  character  |            3980|   3285    |     82.54     |
#> | SecretariatURL  |  character  |            3980|   3285    |     82.54     |
#> |      UNEP       |  character  |            3980|   3778    |     94.92     |
#> |   Supersedes    |  character  |            3980|   3963    |     99.57     |
#> |   References    |  character  |            3980|   3945    |     99.12     |
#> |    EnabledBy    |  character  |            3980|   3873    |     97.31     |
#> |    AmendedBy    |  character  |            3980|   3889    |     97.71     |
#> |       Lit       |  character  |            3980|   3435    |     86.31     |
#> |      Data       |  character  |            3980|   3442    |     86.48     |
#> |      Coded      |  character  |            3980|   3681    |     92.49     |
#> |    Abstract     |  character  |            3980|   3954    |     99.35     |
#> |   TreatyText    |   numeric   |            3980|   2632    |     66.13     |
#> |    Language     |  character  |            3980|   2632    |     66.13     |
#> |Orig_noneng_title|  character  |            3980|   3924    |     98.59     |
#> |      match      |  character  |            3980|   3924    |     98.59     |
#> ------------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> -----------------------------------------------------------------------------
#> |  Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -----------------------------------------------------------------------------
#> |     manyID     |  character  |            3923|     0     |       0       |
#> |    treatyID    |  character  |            3923|     0     |       0       |
#> |     Title      |  character  |            3923|     0     |       0       |
#> |     Begin      |    mdate    |            3923|    47     |      1.2      |
#> |      End       |    mdate    |            3923|   2998    |     76.42     |
#> |   Signature    |    mdate    |            3923|    47     |      1.2      |
#> |     Force      |    mdate    |            3923|   1548    |     39.46     |
#> |  action_area   |  character  |            3923|    452    |     11.52     |
#> |    linkage     |  character  |            3923|   3283    |     83.69     |
#> |   accessionC   |  character  |            3923|    708    |     18.05     |
#> |   accessionP   |  character  |            3923|   3923    |      100      |
#> |termination_type|  character  |            3923|   3828    |     97.58     |
#> |termination_date|  character  |            3923|   3828    |     97.58     |
#> -----------------------------------------------------------------------------
#> 
#> 
#> $IRD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |              92|     0     |       0       |
#> |   treatyID    |  character  |              92|     0     |       0       |
#> |     Title     |  character  |              92|     0     |       0       |
#> |     Begin     |    mdate    |              92|    36     |     39.13     |
#> |RegimeComponent|  character  |              92|    33     |     35.87     |
#> |      Wat      |    mdate    |              92|    34     |     36.96     |
#> |     Wat2      |    mdate    |              92|    78     |     84.78     |
#> |      End      |    mdate    |              92|    36     |     39.13     |
#> | RegimeElement |  character  |              92|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{IEADB: }{
\url https://iea.uoregon.edu/base-agreement-list
}
\item{ECOLEX: }{
\url https://www.ecolex.org/result/?type=treaty
}
\item{CIESIN: }{
\url https://sedac.ciesin.columbia.edu/entri/
}
\item{HEIDI: }{
\url https://www.chaire-epi.ulaval.ca/en/data/heidi
}
\item{HUGGO: }{
\url https://github.com/globalgov/manyenviron
}
\item{AIGGO: }{
\url https://github.com/globalgov/manyenviron
}
\item{IRD: }{
\url https://iea.uoregon.edu/ird
}
}
}

\section{Mapping}{

\itemize{
\item{IEADB: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   'IEA# (click for add'l info)' \tab ieadbID \cr
   'Treaty Name' \tab Title \cr
   'Signature Date' \tab Signature \cr
   'Date IEA entered into force' \tab Force \cr
   'Agreement Type (level 2)' \tab AgreementType \cr
   Inclusion \tab DocType \cr
}


}
\item{ECOLEX: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   EcolexID \tab ecolexID \cr
   title \tab Title \cr
   Date \tab Signature \cr
   'Entry.into.Force' \tab Force \cr
   Document.Type \tab DocType \cr
   Field.of.application \tab GeogArea \cr
}


}
\item{CIESIN: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   'Treaty Title' \tab Title \cr
   'Year of Agreement' \tab Signature \cr
   'Year of Entry into Force' \tab Force \cr
}


}
\item{HEIDI: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   ID \tab heidiID \cr
   'Name.of.the.agreement' \tab Title \cr
   signature.date \tab Signature \cr
}


}
\item{HUGGO: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
\item{AIGGO: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
   } \tab  \cr
   \item{IRD: }{ \tab  \cr
   Variable Mapping \tab  \cr
}
\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Regime \tab Title \cr
   Formation \tab Begin \cr
   Endpoint \tab End \cr
   Watershed \tab Wat \cr
   Watershed2 \tab Wat2 \cr
}


}
}
}

\keyword{datasets}
