% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 5 datasets: IEADB, GNEVAR, ECOLEX, CIESIN, HEIDI.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IEADB: }{A dataset with 3666 observations and the following
10 variables: manyID, Title, Beg, DocType, AgreementType, Signature,
Force, Lineage, treatyID, ieadbID.}
\item{GNEVAR: }{A dataset with 3785 observations and the following
14 variables: manyID, Title, Beg, End, DocType, AgreementType, GeogArea,
Signature, Force, Lineage, Memb.conditions, Memb.procedures, treatyID, gnevarID.}
\item{ECOLEX: }{A dataset with 2174 observations and the following
10 variables: manyID, Title, Beg, DocType, GeogArea, Signature, Force,
Lineage, treatyID, ecolexID.}
\item{CIESIN: }{A dataset with 666 observations and the following
7 variables: manyID, Title, Beg, Signature, Force, Lineage, treatyID.}
\item{HEIDI: }{A dataset with 2280 observations and the following
7 variables: manyID, Title, Beg, Signature, Lineage, treatyID, heidiID.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $IEADB
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            3666|     0     |       0       |
## |     Title     |  character  |            3666|     0     |       0       |
## |      Beg      |    mdate    |            3666|    65     |     1.77      |
## |    DocType    |  character  |            3666|     0     |       0       |
## | AgreementType |  character  |            3666|     0     |       0       |
## |   Signature   |    mdate    |            3666|    65     |     1.77      |
## |     Force     |    mdate    |            3666|    867    |     23.65     |
## |    Lineage    |  character  |            3666|     0     |       0       |
## |   treatyID    |  character  |            3666|     0     |       0       |
## |    ieadbID    |  character  |            3666|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $GNEVAR
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            3785|     0     |       0       |
## |     Title     |  character  |            3785|     0     |       0       |
## |      Beg      |    mdate    |            3785|    57     |     1.51      |
## |      End      |    mdate    |            3785|   3045    |     80.45     |
## |    DocType    |  character  |            3785|    749    |     19.79     |
## | AgreementType |  character  |            3785|    793    |     20.95     |
## |   GeogArea    |  character  |            3785|   1425    |     37.65     |
## |   Signature   |    mdate    |            3785|    57     |     1.51      |
## |     Force     |    mdate    |            3785|   1770    |     46.76     |
## |    Lineage    |  character  |            3785|     0     |       0       |
## |Memb.conditions|  character  |            3785|    683    |     18.04     |
## |Memb.procedures|  character  |            3785|    683    |     18.04     |
## |   treatyID    |  character  |            3785|     0     |       0       |
## |   gnevarID    |  character  |            3785|    753    |     19.89     |
## ----------------------------------------------------------------------------
## 
## 
## $ECOLEX
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            2174|     0     |       0       |
## |     Title     |  character  |            2174|     0     |       0       |
## |      Beg      |    mdate    |            2174|     0     |       0       |
## |    DocType    |  character  |            2174|     0     |       0       |
## |   GeogArea    |  character  |            2174|    90     |     4.14      |
## |   Signature   |    mdate    |            2174|     0     |       0       |
## |     Force     |    mdate    |            2174|   1066    |     49.03     |
## |    Lineage    |  character  |            2174|     0     |       0       |
## |   treatyID    |  character  |            2174|     0     |       0       |
## |   ecolexID    |  character  |            2174|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $CIESIN
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |             666|     0     |       0       |
## |     Title     |  character  |             666|     0     |       0       |
## |      Beg      |    mdate    |             666|     0     |       0       |
## |   Signature   |    mdate    |             666|     0     |       0       |
## |     Force     |    mdate    |             666|     0     |       0       |
## |    Lineage    |  character  |             666|     0     |       0       |
## |   treatyID    |  character  |             666|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $HEIDI
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            2280|     0     |       0       |
## |     Title     |  character  |            2280|     0     |       0       |
## |      Beg      |    mdate    |            2280|    27     |     1.18      |
## |   Signature   |    mdate    |            2280|    27     |     1.18      |
## |    Lineage    |  character  |            2280|     0     |       0       |
## |   treatyID    |  character  |            2280|     0     |       0       |
## |    heidiID    |   numeric   |            2280|     0     |       0       |
## ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
