% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 4 datasets: ECOLEX_MEM, GNEVAR_MEM, IEADB_MEM, TFDD_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ECOLEX_MEM: }{A dataset with 25003 observations and the following
10 variables: manyID, CountryID, Title, Beg, End, SignatureCountry, Force,
Rat, ecolexID, treatyID.}
\item{GNEVAR_MEM: }{A dataset with 35671 observations and the following
13 variables: manyID, CountryID, Title, Beg, End, SignatureCountry,
Signature, Rat, Force, Term, Withdrawal, gnevarID, treatyID.}
\item{IEADB_MEM: }{A dataset with 15466 observations and the following
12 variables: manyID, CountryID, Title, Beg, End, SignatureCountry,
Signature, Rat, Force, DocType, ieadbID, treatyID.}
\item{TFDD_MEM: }{A dataset with 2118 observations and the following
8 variables: manyID, CountryID, Title, Beg, Signature, tfddID,
Memberships, treatyID.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\preformatted{## $ECOLEX_MEM
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             25003 
## Number of columns          10    
## _______________________          
## Column type frequency:           
##   character                5     
##   messydt                  5     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0         1         9    41     0     5956          0
## 2 CountryID            22         0.999     3     3     0      214          0
## 3 Title                 0         1        15   492     0     2141          0
## 4 ecolexID              0         1        10    10     0     2169          0
## 5 treatyID              0         1         9    41     0     6145          0
## 
## ── Variable type: messydt ──────────────────────────────────────────────────────
##   skim_variable    n_missing complete_rate n_empty n_unique
## 1 Beg                   3447        0.862        0     7510
## 2 End                  24539        0.0186       0      271
## 3 SignatureCountry     15339        0.387        0     2630
## 4 Force                 5809        0.768        0     7292
## 5 Rat                   7021        0.719        0     8383
## 
## $GNEVAR_MEM
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             35671 
## Number of columns          13    
## _______________________          
## Column type frequency:           
##   character                5     
##   messydt                  8     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0         1         9    41     0     2288          0
## 2 CountryID             0         1         3    40     0      239          0
## 3 Title               221         0.994    15   550     0     2167          0
## 4 gnevarID             17         1.00      9     9     0     2322          0
## 5 treatyID              0         1         9    41     0     2307          0
## 
## ── Variable type: messydt ──────────────────────────────────────────────────────
##   skim_variable    n_missing complete_rate n_empty n_unique
## 1 Beg                     15        1.00         0     1974
## 2 End                  31818        0.108        0      663
## 3 SignatureCountry        46        0.999        0     1947
## 4 Signature               46        0.999        0     1947
## 5 Rat                  19419        0.456        0     7221
## 6 Force                 4359        0.878        0     6751
## 7 Term                 34514        0.0324       0      287
## 8 Withdrawal           31968        0.104        0      660
## 
## $IEADB_MEM
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             15466 
## Number of columns          12    
## _______________________          
## Column type frequency:           
##   character                5     
##   messydt                  6     
##   numeric                  1     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0             1     9    41     0     4390          0
## 2 CountryID             0             1     3     7     0      196          0
## 3 Title                 0             1    16   517     0     2153          0
## 4 DocType               0             1     1     1     0        2          0
## 5 treatyID              0             1     9    41     0     4427          0
## 
## ── Variable type: messydt ──────────────────────────────────────────────────────
##   skim_variable    n_missing complete_rate n_empty n_unique
## 1 Beg                      0        1            0     6312
## 2 End                  14595        0.0563       0       70
## 3 SignatureCountry      5799        0.625        0     3154
## 4 Signature                0        1            0     1933
## 5 Rat                   5486        0.645        0     6426
## 6 Force                    0        1            0     6607
## 
## ── Variable type: numeric ──────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate  mean    sd    p0   p25   p50   p75
## 1 ieadbID               0             1 3232. 1627.    32  2702  3003  3346
##    p100
## 1  8919
## 
## $TFDD_MEM
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             2118  
## Number of columns          8     
## _______________________          
## Column type frequency:           
##   character                5     
##   messydt                  2     
##   numeric                  1     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0         1         6    41     0      627          0
## 2 CountryID             2         0.999     3     3     0      158          0
## 3 Title                 2         0.999    17   455     0      630          0
## 4 Memberships          28         0.987     3   622     0      239          0
## 5 treatyID              0         1         6    41     0      634          0
## 
## ── Variable type: messydt ──────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique
## 1 Beg                 133         0.937       0      553
## 2 Signature           133         0.937       0      553
## 
## ── Variable type: numeric ──────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate  mean    sd    p0   p25   p50   p75
## 1 tfddID                2         0.999  331.  189.     1   174   329   497
##    p100
## 1   646
}
}
\keyword{datasets}
