% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 4 datasets: ECOLEX_MEM, GNEVAR_MEM, IEADB_MEM, TFDD_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ECOLEX_MEM: }{A dataset with 25003 observations and the following
10 variables: manyID, CountryID, Title, Beg, End, SignatureC, Force, Rat,
ecolexID, treatyID.}
\item{GNEVAR_MEM: }{A dataset with 35671 observations and the following
13 variables: manyID, CountryID, Title, Beg, End, SignatureC, Signature,
Rat, Force, Term, Withdrawal, gnevarID, treatyID.}
\item{IEADB_MEM: }{A dataset with 15466 observations and the following
12 variables: manyID, CountryID, Title, Beg, End, SignatureC, Signature,
Rat, Force, L, ieadbID, treatyID.}
\item{TFDD_MEM: }{A dataset with 2118 observations and the following
8 variables: manyID, CountryID, Title, Beg, Signature, tfddID,
Memberships, treatyID.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
