% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships datacube documentation}
\format{
The memberships datacube is a list that contains the
following 4 datasets: ECOLEX_MEM, IEADB_MEM, TFDD_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{ECOLEX_MEM: }{A dataset with 25003 observations and the following
10 variables: manyID, stateID, Title, Begin, End, stateSignature, Force, Rat,
ecolexID, treatyID.}
\item{IEADB_MEM: }{A dataset with 15463 observations and the following
12 variables: manyID, stateID, Title, Begin, End, stateSignature, Signature,
Rat, Force, DocType, ieadbID, treatyID.}
\item{TFDD_MEM: }{A dataset with 2118 observations and the following
8 variables: manyID, stateID, Title, Begin, Signature, tfddID, Memberships,
treatyID.}
\item{HUGGO_MEM: }{A dataset with 51042 observations and the following
31 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force, End,
StateSignature, StateRatification, StateForce, StateForce2, StateForce3,
StateEnd, StateEnd2, gengID, ieaID, ecolexID, Accession, Succession, Term,
Comments, Deposit, obsolete, ProvisionalApp, Reservation, Notes,
stateForce_ecolex, stateForce_iea, Consent, Acceptance.}
}
}
\source{
\itemize{
\item{ECOLEX_MEM: }{
J. Sommer. “Global governance in forestry: a cross-national analysis”. In: \emph{International Journal ofSustainable Development & World Ecology} 27.6 (2020), pp. 481-495.}
\item{IEADB_MEM: }{
R. B. Mitchell, L. B. Andonova, M. Axelrod, et al. “What we know (and could know) about internationalenvironmental agreements”. In: \emph{Global Environmental Politics} 20.1 (2020), pp. 103-121.}
\item{TFDD_MEM: }{
O. College of Earth and O. S. U. Atmospheric Science. \emph{Product of the Transboundary Freshwater DisputeDatabase}. \url{http://transboundarywaters.science.oregonstate.edu.}. 2021.}
\item{HUGGO_MEM: }{
NA}
}
}
\usage{
memberships
}
\description{
memberships datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $ECOLEX_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           25003|   3447    |     13.79     |
#> |    stateID    |  character  |           25003|    22     |     0.09      |
#> |     Title     |  character  |           25003|     0     |       0       |
#> |     Begin     |    mdate    |           25003|   3447    |     13.79     |
#> |      End      |    mdate    |           25003|   24539   |     98.14     |
#> |stateSignature |    mdate    |           25003|   15339   |     61.35     |
#> |     Force     |    mdate    |           25003|   5809    |     23.23     |
#> |      Rat      |    mdate    |           25003|   7021    |     28.08     |
#> |   ecolexID    |  character  |           25003|     0     |       0       |
#> |   treatyID    |  character  |           25003|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IEADB_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           15463|     0     |       0       |
#> |    stateID    |  character  |           15463|     0     |       0       |
#> |     Title     |  character  |           15463|     0     |       0       |
#> |     Begin     |    mdate    |           15463|     0     |       0       |
#> |      End      |    mdate    |           15463|   14593   |     94.37     |
#> |stateSignature |    mdate    |           15463|   5797    |     37.49     |
#> |   Signature   |    mdate    |           15463|     0     |       0       |
#> |      Rat      |    mdate    |           15463|   5485    |     35.47     |
#> |     Force     |    mdate    |           15463|     0     |       0       |
#> |    DocType    |  character  |           15463|     0     |       0       |
#> |    ieadbID    |   numeric   |           15463|     0     |       0       |
#> |   treatyID    |  character  |           15463|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TFDD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2118|    133    |     6.28      |
#> |    stateID    |  character  |            2118|     2     |     0.09      |
#> |     Title     |  character  |            2118|     2     |     0.09      |
#> |     Begin     |    mdate    |            2118|    133    |     6.28      |
#> |   Signature   |    mdate    |            2118|    133    |     6.28      |
#> |    tfddID     |   numeric   |            2118|     2     |     0.09      |
#> |  Memberships  |  character  |            2118|    843    |     39.8      |
#> |   treatyID    |  character  |            2118|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           51042|     0     |       0       |
#> |    treatyID     |  character  |           51042|     0     |       0       |
#> |      Title      |  character  |           51042|     0     |       0       |
#> |      Begin      |    mdate    |           51042|    12     |     0.02      |
#> |     stateID     |  character  |           51042|     0     |       0       |
#> |    Signature    |    mdate    |           51042|    14     |     0.03      |
#> |      Force      |    mdate    |           51042|   3537    |     6.93      |
#> |       End       |    mdate    |           51042|   38689   |     75.8      |
#> | StateSignature  |    mdate    |           51042|   11889   |     23.29     |
#> |StateRatification|    mdate    |           51042|   27634   |     54.14     |
#> |   StateForce    |    mdate    |           51042|   18871   |     36.97     |
#> |   StateForce2   |    mdate    |           51042|   44319   |     86.83     |
#> |   StateForce3   |    mdate    |           51042|   47337   |     92.74     |
#> |    StateEnd     |    mdate    |           51042|   32876   |     64.41     |
#> |    StateEnd2    |    mdate    |           51042|   47331   |     92.73     |
#> |     gengID      |  character  |           51042|     0     |       0       |
#> |      ieaID      |  character  |           51042|   8629    |     16.91     |
#> |    ecolexID     |  character  |           51042|   30963   |     60.66     |
#> |    Accession    |    mdate    |           51042|   43027   |     84.3      |
#> |   Succession    |   numeric   |           51042|   46987   |     92.06     |
#> |      Term       |    mdate    |           51042|   49864   |     97.69     |
#> |    Comments     |  character  |           51042|   47760   |     93.57     |
#> |     Deposit     |  character  |           51042|   49385   |     96.75     |
#> |    obsolete     |  character  |           51042|   38414   |     75.26     |
#> | ProvisionalApp  |  character  |           51042|   50268   |     98.48     |
#> |   Reservation   |  character  |           51042|   50268   |     98.48     |
#> |      Notes      |  character  |           51042|   50856   |     99.64     |
#> |stateForce_ecolex|    mdate    |           51042|   32518   |     63.71     |
#> | stateForce_iea  |    mdate    |           51042|   32518   |     63.71     |
#> |     Consent     |  character  |           51042|   50817   |     99.56     |
#> |   Acceptance    |  character  |           51042|   50515   |     98.97     |
#> ------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{ECOLEX_MEM: }{
\url{https://www.ecolex.org/result/?type=treaty}
}
\item{IEADB_MEM: }{
\url{https://iea.uoregon.edu/country-members}
}
\item{TFDD_MEM: }{
\url{https://transboundarywaters.science.oregonstate.edu/}
}
\item{HUGGO_MEM: }{Hand-coded data by the GGO team.}
}
}

\section{Mapping}{

\itemize{
\item{ECOLEX_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   For \tab Force \cr
   Rati \tab Rat \cr
   Sign \tab stateSignature \cr
   Term \tab End \cr
   StatID \tab stateID \cr
   EcolexID \tab ecolexID \cr
}


}
\item{IEADB_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   country \tab stateID \cr
   treatyname \tab Title \cr
   tsig \tab Signature \cr
   csig \tab stateSignature \cr
   crat \tab Rat \cr
   tterm \tab End \cr
   ceif3 \tab Force \cr
   ceif4 \tab Force2 \cr
   mitch_id \tab ieadbID \cr
   inclusion \tab DocType \cr
}


}
\item{TFDD_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   DateSigned \tab Signature \cr
   '2016Update ID' \tab tfddID \cr
   CCODE \tab stateID \cr
   DocumentName \tab Title \cr
   Signatories \tab Memberships \cr
}


}
}
}

\keyword{datasets}
