% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qData-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the following 3 datasets: IEA_MEM, GNEVAR_MEM, ECOLEX_MEM.
For more information and references to each of the datasets used, please use the report_data() function.
\describe{
\item{IEA_MEM: }{A dataset with 27410 observations and the following 5 variables: ID, Country, Title, Beg, End.}
\item{GNEVAR_MEM: }{A dataset with 35671 observations and the following 5 variables: ID, Country, Title, Beg, End.}
\item{ECOLEX_MEM: }{A dataset with 25003 observations and the following 4 variables: ID, Country, Beg, End.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\keyword{datasets}
