% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 4 datasets: ECOLEX_MEM, IEADB_MEM, TFDD_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ECOLEX_MEM: }{A dataset with 25003 observations and the following
10 variables: manyID, stateID, Title, Begin, End, stateSignature, Force, Rat,
ecolexID, treatyID.}
\item{IEADB_MEM: }{A dataset with 15463 observations and the following
12 variables: manyID, stateID, Title, Begin, End, stateSignature, Signature,
Rat, Force, DocType, ieadbID, treatyID.}
\item{TFDD_MEM: }{A dataset with 2118 observations and the following
8 variables: manyID, stateID, Title, Begin, Signature, tfddID, Memberships,
treatyID.}
\item{HUGGO_MEM: }{A dataset with 51740 observations and the following
31 variables: manyID, treatyID, Title, Begin, stateID, End, Signature, Force,
stateSignature, stateRat, stateForce, stateForce2, stateForce3, Term,
stateWithdrawal, stateWithdrawal2, gengID, ecolexID, ieaID, Comments,
Deposit, obsolete, ProvisionalApp, Reservation, verified, Notes,
stateForce_ecolex, stateForce_iea, Consent, Acceptance, Accession.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $ECOLEX_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           25003|   3447    |     13.79     |
#> |    stateID    |  character  |           25003|    22     |     0.09      |
#> |     Title     |  character  |           25003|     0     |       0       |
#> |     Begin     |    mdate    |           25003|   3447    |     13.79     |
#> |      End      |    mdate    |           25003|   24539   |     98.14     |
#> |stateSignature |    mdate    |           25003|   15339   |     61.35     |
#> |     Force     |    mdate    |           25003|   5809    |     23.23     |
#> |      Rat      |    mdate    |           25003|   7021    |     28.08     |
#> |   ecolexID    |  character  |           25003|     0     |       0       |
#> |   treatyID    |  character  |           25003|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IEADB_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           15463|     0     |       0       |
#> |    stateID    |  character  |           15463|     0     |       0       |
#> |     Title     |  character  |           15463|     0     |       0       |
#> |     Begin     |    mdate    |           15463|     0     |       0       |
#> |      End      |    mdate    |           15463|   14593   |     94.37     |
#> |stateSignature |    mdate    |           15463|   5797    |     37.49     |
#> |   Signature   |    mdate    |           15463|     0     |       0       |
#> |      Rat      |    mdate    |           15463|   5485    |     35.47     |
#> |     Force     |    mdate    |           15463|     0     |       0       |
#> |    DocType    |  character  |           15463|     0     |       0       |
#> |    ieadbID    |   numeric   |           15463|     0     |       0       |
#> |   treatyID    |  character  |           15463|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TFDD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2118|    133    |     6.28      |
#> |    stateID    |  character  |            2118|     2     |     0.09      |
#> |     Title     |  character  |            2118|     2     |     0.09      |
#> |     Begin     |    mdate    |            2118|    133    |     6.28      |
#> |   Signature   |    mdate    |            2118|    133    |     6.28      |
#> |    tfddID     |   numeric   |            2118|     2     |     0.09      |
#> |  Memberships  |  character  |            2118|    843    |     39.8      |
#> |   treatyID    |  character  |            2118|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           51740|     0     |       0       |
#> |    treatyID     |  character  |           51740|     0     |       0       |
#> |      Title      |  character  |           51740|     0     |       0       |
#> |      Begin      |    mdate    |           51740|    12     |     0.02      |
#> |     stateID     |  character  |           51740|     0     |       0       |
#> |       End       |    mdate    |           51740|   39723   |     76.77     |
#> |    Signature    |    mdate    |           51740|    14     |     0.03      |
#> |      Force      |    mdate    |           51740|   3529    |     6.82      |
#> | stateSignature  |    mdate    |           51740|   11488   |     22.2      |
#> |    stateRat     |    mdate    |           51740|   27610   |     53.36     |
#> |   stateForce    |    mdate    |           51740|   21359   |     41.28     |
#> |   stateForce2   |    mdate    |           51740|   45130   |     87.22     |
#> |   stateForce3   |    mdate    |           51740|   48021   |     92.81     |
#> |      Term       |  character  |           51740|   50566   |     97.73     |
#> | stateWithdrawal |    mdate    |           51740|   38725   |     74.85     |
#> |stateWithdrawal2 |    mdate    |           51740|   48015   |     92.8      |
#> |     gengID      |  character  |           51740|     0     |       0       |
#> |    ecolexID     |  character  |           51740|   30677   |     59.29     |
#> |      ieaID      |  character  |           51740|   8175    |     15.8      |
#> |    Comments     |  character  |           51740|   48481   |     93.7      |
#> |     Deposit     |  character  |           51740|   49900   |     96.44     |
#> |    obsolete     |  character  |           51740|   38703   |     74.8      |
#> | ProvisionalApp  |  character  |           51740|   50926   |     98.43     |
#> |   Reservation   |  character  |           51740|   50936   |     98.45     |
#> |    verified     |  character  |           51740|   48516   |     93.77     |
#> |      Notes      |  character  |           51740|   51554   |     99.64     |
#> |stateForce_ecolex|  character  |           51740|   32357   |     62.54     |
#> | stateForce_iea  |  character  |           51740|   32357   |     62.54     |
#> |     Consent     |  character  |           51740|   51441   |     99.42     |
#> |   Acceptance    |  character  |           51740|   51651   |     99.83     |
#> |    Accession    |  character  |           51740|   45685   |     88.3      |
#> ------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
