% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qData-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the following  datasets: IEA_MEM, GNEVAR_MEM, ECOLEX_MEM. For more info, please use the report_data() function.
\describe{
\item{IEA_MEM: A dataset with 27410 observations and the following 5 variables: ID, IEA_ID, Title, Beg, End.}{Describe dataset here}
\item{GNEVAR_MEM: A dataset with 35671 observations and the following 5 variables: ID, GNEVAR_ID, Title, Beg, End.}{Describe dataset here}
\item{ECOLEX_MEM: A dataset with 25003 observations and the following 4 variables: ID, ECOLEX_ID, Beg, End.}{Describe dataset here}
}
}
\usage{
memberships
}
\description{
Write the database citation here.
}
\keyword{datasets}
