% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qData-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 4 datasets: ECOLEX_MEM, GNEVAR_MEM, IEADB_MEM, TFDD_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{ECOLEX_MEM: }{A dataset with 25003 observations and the following 9 variables: ECOLEX_ID, Title, Country, Beg, End, SignatureC, Force, Rat, qID.}
\item{GNEVAR_MEM: }{A dataset with 35671 observations and the following 12 variables: GNEVAR_ID, Country, Title, Beg, End, SignatureC, Signature, Rat, Force, Term, Withdrawal, qID.}
\item{IEADB_MEM: }{A dataset with 27410 observations and the following 10 variables: IEADB_ID, Country, Title, Beg, End, SignatureC, Signature, Rat, Force, qID.}
\item{TFDD_MEM: }{A dataset with 2118 observations and the following 6 variables: TFDD_ID, Country, Title, Beg, Signature, qID.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\keyword{datasets}
