% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qData-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the following 4 datasets: IEA_MEM, GNEVAR_MEM, ECOLEX_MEM, TFDD_MEM.
For more information and references to each of the datasets used, please use the report_data() function.
\describe{
\item{IEA_MEM: }{A dataset with 27410 observations and the following 9 variables: IEADB_ID, Country, Title, Beg, End, SignatureC, Signature, Rat, Force.}
\item{GNEVAR_MEM: }{A dataset with 35671 observations and the following 11 variables: GNEVAR_ID, Country, Title, Beg, End, SignatureC, Signature, Rat, Force, Term, Withdrawal.}
\item{ECOLEX_MEM: }{A dataset with 25003 observations and the following 7 variables: ECOLEX_ID, Country, Beg, End, SignatureC, Force, Rat.}
\item{TFDD_MEM: }{A dataset with 2118 observations and the following 4 variables: TFDD_ID, Country, Title, Signature.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\keyword{datasets}
