% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 4 datasets: ECOLEX_MEM, IEADB_MEM, TFDD_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{ECOLEX_MEM: }{A dataset with 24975 observations and the following
10 variables: manyID, treatyID, Title, Beg, stateID, End, stateSignature,
Force, Rat, ecolexID.}
\item{IEADB_MEM: }{A dataset with 15463 observations and the following
12 variables: manyID, treatyID, Title, Beg, stateID, End, stateSignature,
Signature, Rat, Force, DocType, ieadbID.}
\item{TFDD_MEM: }{A dataset with 1832 observations and the following
8 variables: manyID, treatyID, Title, Beg, stateID, Signature, tfddID,
Memberships.}
\item{HUGGO_MEM: }{A dataset with 52462 observations and the following
31 variables: manyID, treatyID, Title, Beg, stateID, End, Signature, Force,
stateSignature, stateRat, stateForce, stateForce2, stateForce3, Term,
stateWithdrawal, stateWithdrawal2, gengID, ecolexID, ieaID, Comments,
Deposit, obsolete, ProvisionalApp, Reservation, verified, Notes,
stateForce_ecolex, stateForce_iea, Consent, Acceptance, Accession.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $ECOLEX_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           24975|     0     |       0       |
#> |   treatyID    |  character  |           24975|     0     |       0       |
#> |     Title     |  character  |           24975|     0     |       0       |
#> |      Beg      |    mdate    |           24975|   3421    |     13.7      |
#> |    stateID    |  character  |           24975|    21     |     0.08      |
#> |      End      |    mdate    |           24975|   24511   |     98.14     |
#> |stateSignature |    mdate    |           24975|   15313   |     61.31     |
#> |     Force     |    mdate    |           24975|   5783    |     23.16     |
#> |      Rat      |    mdate    |           24975|   6993    |      28       |
#> |   ecolexID    |  character  |           24975|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IEADB_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           15463|     0     |       0       |
#> |   treatyID    |  character  |           15463|     0     |       0       |
#> |     Title     |  character  |           15463|     0     |       0       |
#> |      Beg      |    mdate    |           15463|     0     |       0       |
#> |    stateID    |  character  |           15463|     0     |       0       |
#> |      End      |    mdate    |           15463|   14593   |     94.37     |
#> |stateSignature |    mdate    |           15463|   5797    |     37.49     |
#> |   Signature   |    mdate    |           15463|     0     |       0       |
#> |      Rat      |    mdate    |           15463|   5485    |     35.47     |
#> |     Force     |    mdate    |           15463|     0     |       0       |
#> |    DocType    |  character  |           15463|     0     |       0       |
#> |    ieadbID    |   numeric   |           15463|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TFDD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            1832|     0     |       0       |
#> |   treatyID    |  character  |            1832|     0     |       0       |
#> |     Title     |  character  |            1832|     1     |     0.05      |
#> |      Beg      |    mdate    |            1832|    102    |     5.57      |
#> |    stateID    |  character  |            1832|     1     |     0.05      |
#> |   Signature   |    mdate    |            1832|    102    |     5.57      |
#> |    tfddID     |   numeric   |            1832|     1     |     0.05      |
#> |  Memberships  |  character  |            1832|    805    |     43.94     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           52462|     0     |       0       |
#> |    treatyID     |  character  |           52462|     0     |       0       |
#> |      Title      |  character  |           52462|     0     |       0       |
#> |       Beg       |    mdate    |           52462|    12     |     0.02      |
#> |     stateID     |  character  |           52462|     0     |       0       |
#> |       End       |    mdate    |           52462|   40433   |     77.07     |
#> |    Signature    |    mdate    |           52462|    14     |     0.03      |
#> |      Force      |    mdate    |           52462|   3538    |     6.74      |
#> | stateSignature  |  character  |           52462|   11488   |     21.9      |
#> |    stateRat     |  character  |           52462|   27944   |     53.27     |
#> |   stateForce    |  character  |           52462|   21420   |     40.83     |
#> |   stateForce2   |  character  |           52462|   45852   |     87.4      |
#> |   stateForce3   |  character  |           52462|   48743   |     92.91     |
#> |      Term       |  character  |           52462|   51288   |     97.76     |
#> | stateWithdrawal |  character  |           52462|   39438   |     75.17     |
#> |stateWithdrawal2 |  character  |           52462|   48737   |     92.9      |
#> |     gengID      |  character  |           52462|     0     |       0       |
#> |    ecolexID     |  character  |           52462|   31035   |     59.16     |
#> |      ieaID      |  character  |           52462|   8469    |     16.14     |
#> |    Comments     |  character  |           52462|   49199   |     93.78     |
#> |     Deposit     |  character  |           52462|   50622   |     96.49     |
#> |    obsolete     |  character  |           52462|   39300   |     74.91     |
#> | ProvisionalApp  |  character  |           52462|   51648   |     98.45     |
#> |   Reservation   |  character  |           52462|   51658   |     98.47     |
#> |    verified     |  character  |           52462|   49234   |     93.85     |
#> |      Notes      |  character  |           52462|   52276   |     99.65     |
#> |stateForce_ecolex|  character  |           52462|   33011   |     62.92     |
#> | stateForce_iea  |  character  |           52462|   33011   |     62.92     |
#> |     Consent     |  character  |           52462|   52163   |     99.43     |
#> |   Acceptance    |  character  |           52462|   52373   |     99.83     |
#> |    Accession    |  character  |           52462|   46400   |     88.44     |
#> ------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
