% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 4 datasets: ECOLEX_MEM, IEADB_MEM, TFDD_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ECOLEX_MEM: }{A dataset with 25003 observations and the following
10 variables: manyID, stateID, Title, Beg, End, stateSignature, Force, Rat,
ecolexID, treatyID.}
\item{IEADB_MEM: }{A dataset with 15466 observations and the following
12 variables: manyID, stateID, Title, Beg, End, stateSignature, Signature,
Rat, Force, DocType, ieadbID, treatyID.}
\item{TFDD_MEM: }{A dataset with 2118 observations and the following
8 variables: manyID, stateID, Title, Beg, Signature, tfddID, Memberships,
treatyID.}
\item{HUGGO_MEM: }{A dataset with 52462 observations and the following
31 variables: manyID, treatyID, stateID, Title, Beg, End, Signature, Force,
stateSignature, stateRat, stateForce, stateForce2, stateForce3, Term,
stateWithdrawal, stateWithdrawal2, gengID, ecolexID, ieaID, Comments,
Deposit, obsolete, ProvisionalApp, Reservation, verified, Notes,
stateForce_ecolex, stateForce_iea, Consent, Acceptance, Accession.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $ECOLEX_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           25003|     0     |       0       |
#> |    stateID    |  character  |           25003|    22     |     0.09      |
#> |     Title     |  character  |           25003|     0     |       0       |
#> |      Beg      |    mdate    |           25003|   3447    |     13.79     |
#> |      End      |    mdate    |           25003|   24539   |     98.14     |
#> |stateSignature |    mdate    |           25003|   15339   |     61.35     |
#> |     Force     |    mdate    |           25003|   5809    |     23.23     |
#> |      Rat      |    mdate    |           25003|   7021    |     28.08     |
#> |   ecolexID    |  character  |           25003|     0     |       0       |
#> |   treatyID    |  character  |           25003|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IEADB_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           15466|     0     |       0       |
#> |    stateID    |  character  |           15466|     0     |       0       |
#> |     Title     |  character  |           15466|     0     |       0       |
#> |      Beg      |    mdate    |           15466|     0     |       0       |
#> |      End      |    mdate    |           15466|   14595   |     94.37     |
#> |stateSignature |    mdate    |           15466|   5799    |     37.5      |
#> |   Signature   |    mdate    |           15466|     0     |       0       |
#> |      Rat      |    mdate    |           15466|   5486    |     35.47     |
#> |     Force     |    mdate    |           15466|     0     |       0       |
#> |    DocType    |  character  |           15466|     0     |       0       |
#> |    ieadbID    |   numeric   |           15466|     0     |       0       |
#> |   treatyID    |  character  |           15466|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TFDD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2118|     0     |       0       |
#> |    stateID    |  character  |            2118|     2     |     0.09      |
#> |     Title     |  character  |            2118|     2     |     0.09      |
#> |      Beg      |    mdate    |            2118|    133    |     6.28      |
#> |   Signature   |    mdate    |            2118|    133    |     6.28      |
#> |    tfddID     |   numeric   |            2118|     2     |     0.09      |
#> |  Memberships  |  character  |            2118|    843    |     39.8      |
#> |   treatyID    |  character  |            2118|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           52462|     0     |       0       |
#> |    treatyID     |  character  |           52462|     0     |       0       |
#> |     stateID     |  character  |           52462|     0     |       0       |
#> |      Title      |  character  |           52462|     0     |       0       |
#> |       Beg       |    mdate    |           52462|    12     |     0.02      |
#> |       End       |    mdate    |           52462|   40433   |     77.07     |
#> |    Signature    |    mdate    |           52462|    14     |     0.03      |
#> |      Force      |    mdate    |           52462|   3538    |     6.74      |
#> | stateSignature  |  character  |           52462|   11488   |     21.9      |
#> |    stateRat     |  character  |           52462|   27944   |     53.27     |
#> |   stateForce    |  character  |           52462|   21420   |     40.83     |
#> |   stateForce2   |  character  |           52462|   45852   |     87.4      |
#> |   stateForce3   |  character  |           52462|   48743   |     92.91     |
#> |      Term       |  character  |           52462|   51288   |     97.76     |
#> | stateWithdrawal |  character  |           52462|   39438   |     75.17     |
#> |stateWithdrawal2 |  character  |           52462|   48737   |     92.9      |
#> |     gengID      |  character  |           52462|     0     |       0       |
#> |    ecolexID     |  character  |           52462|   31035   |     59.16     |
#> |      ieaID      |  character  |           52462|   8469    |     16.14     |
#> |    Comments     |  character  |           52462|   49199   |     93.78     |
#> |     Deposit     |  character  |           52462|   50622   |     96.49     |
#> |    obsolete     |  character  |           52462|   39300   |     74.91     |
#> | ProvisionalApp  |  character  |           52462|   51648   |     98.45     |
#> |   Reservation   |  character  |           52462|   51658   |     98.47     |
#> |    verified     |  character  |           52462|   49234   |     93.85     |
#> |      Notes      |  character  |           52462|   52276   |     99.65     |
#> |stateForce_ecolex|  character  |           52462|   33011   |     62.92     |
#> | stateForce_iea  |  character  |           52462|   33011   |     62.92     |
#> |     Consent     |  character  |           52462|   52163   |     99.43     |
#> |   Acceptance    |  character  |           52462|   52373   |     99.83     |
#> |    Accession    |  character  |           52462|   46400   |     88.44     |
#> ------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
