% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qEnviron-organizations.R
\docType{data}
\name{organizations}
\alias{organizations}
\title{organizations database documentation}
\format{
The organizations database is a list that contains the
following 1 datasets: MIA.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{MIA: }{A dataset with 78 observations and the following
4 variables: Name, Beg, End, COW.}
}
}
\usage{
organizations
}
\description{
organizations database documentation
}
\keyword{datasets}
