% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-organizations.R
\docType{data}
\name{organizations}
\alias{organizations}
\title{organizations database documentation}
\format{
The organizations database is a list that contains the
following 2 datasets: MIA, YIO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{MIA: }{A dataset with 78 observations and the following
4 variables: Name, Beg, End, COW.}
\item{YIO: }{A dataset with 75115 observations and the following
4 variables: Title, Abbreviation, Beg, Country.}
}
}
\usage{
organizations
}
\description{
organizations database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $MIA
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |     Name      |  character  |              78|     0     |       0       |
## |      Beg      |    mdate    |              78|     0     |       0       |
## |      End      |    mdate    |              78|    78     |      100      |
## |      COW      |   numeric   |              78|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $YIO
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |     Title     |  character  |           75115|     0     |       0       |
## | Abbreviation  |  character  |           75115|   29296   |      39       |
## |      Beg      |    mdate    |           75115|   23141   |     30.81     |
## |    Country    |  character  |           75115|   28101   |     37.41     |
## ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
