% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qData-references.R, R/qEnviron-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 1 datasets: REF.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{REF: }{A dataset with 1164 observations and the following 3 variables: qID1, RefType, qID2.}
}

The references database is a list that contains the
following 2 datasets: ECOLEX, ECOLEX_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{ECOLEX: }{A dataset with 1164 observations and the following
3 variables: qID1, RefType, qID2.}
\item{ECOLEX_REF: }{A dataset with 1164 observations and the following
3 variables: Treaty1, RefType, Treaty2.}
}
}
\usage{
references

references
}
\description{
references database documentation

references database documentation
}
\keyword{datasets}
