% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qEnviron-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 1 datasets: REF.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{REF: }{A dataset with 2174 observations and the following
3 variables: Treaty1, Treaty2, ReferenceType.}
}
}
\usage{
references
}
\description{
references database documentation
}
\keyword{datasets}
