% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-references.R
\docType{data}
\name{references}
\alias{references}
\title{references datacube documentation}
\format{
The references datacube is a list that contains the
following 1 datasets: ECOLEX_REF.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{ECOLEX_REF: }{A dataset with 1164 observations and the following
3 variables: treatyID1, treatyID2, RefType.}
}
}
\source{
\itemize{
\item{ECOLEX_REF: }{
R. E. Kim and B. Mackey. “International environmental law as a complex adaptive system”. In: \emph{International Environmental Agreements: Politics, Law and Economics} 14.1 (2014), pp. 5-24.}
}
}
\usage{
references
}
\description{
references datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $ECOLEX_REF
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |   treatyID1   |  character  |            1164|     0     |       0       |
#> |   treatyID2   |  character  |            1164|     0     |       0       |
#> |    RefType    |  character  |            1164|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{ECOLEX_REF: }{NA}
}
}

\section{Mapping}{

\itemize{
\item{ECOLEX_REF: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
}
}

\keyword{datasets}
