% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 1 datasets: ECOLEX_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ECOLEX_REF: }{A dataset with 1164 observations and the following
3 variables: treatyID1, treatyID2, RefType.}
}
}
\usage{
references
}
\description{
references database documentation
}
\details{
\preformatted{## $ECOLEX_REF
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             1164  
## Number of columns          3     
## _______________________          
## Column type frequency:           
##   character                3     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 treatyID1             0             1     9    40     0      648          0
## 2 treatyID2             0             1     9    40     0      648          0
## 3 RefType               0             1     5    13     0        8          0
}
}
\keyword{datasets}
