% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-regimes.R
\docType{data}
\name{regimes}
\alias{regimes}
\title{regimes datacube documentation}
\format{
The regimes datacube is a list that contains the
following 1 datasets: IRD.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IRD: }{A dataset with 92 observations and the following
7 variables: Regime, RegimeComponent, Begin, Wat, Wat2, End, RegimeElement.}
}
}
\source{
\itemize{
\item{IRD: }{
\link{1} O. R. Young and M. Zürn. “The international regimes database: Designing and using a sophisticatedtool for institutional analysis”. In: \emph{Global Environmental Politics} 6.3 (2006), pp. 121-143.}
}
}
\usage{
regimes
}
\description{
regimes datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $IRD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    Regime     |  character  |              92|     0     |       0       |
#> |RegimeComponent|  character  |              92|    33     |     35.87     |
#> |     Begin     |    mdate    |              92|    36     |     39.13     |
#> |      Wat      |   messydt   |              92|    34     |     36.96     |
#> |     Wat2      |   messydt   |              92|    78     |     84.78     |
#> |      End      |    mdate    |              92|    36     |     39.13     |
#> | RegimeElement |  character  |              92|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{IRD: }{ \url https://direct.mit.edu/glep/article-abstract/6/3/121/14360/The-International-Regimes-Database-Designing-and}
}
}

\section{Mapping}{

\itemize{
\item{IRD: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Formation \tab Begin \cr
   Endpoint \tab End \cr
   Watershed \tab Wat \cr
   Watershed2 \tab Wat2 \cr
}


}
}
}

\keyword{datasets}
