% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-regimes.R
\docType{data}
\name{regimes}
\alias{regimes}
\title{regimes database documentation}
\format{
The regimes database is a list that contains the
following 1 datasets: IRD.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{IRD: }{A dataset with 92 observations and the following
7 variables: Regime, RegimeComponent, Beg, Wat, Wat2, End, RegimeElement.}
}
}
\usage{
regimes
}
\description{
regimes database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $IRD
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    Regime     |  character  |              92|     0     |       0       |
## |RegimeComponent|  character  |              92|    33     |     35.87     |
## |      Beg      |    mdate    |              92|    36     |     39.13     |
## |      Wat      |   messydt   |              92|    34     |     36.96     |
## |     Wat2      |   messydt   |              92|    78     |     84.78     |
## |      End      |    mdate    |              92|    36     |     39.13     |
## | RegimeElement |  character  |              92|     0     |       0       |
## ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
