% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 1 datasets: GNEVAR_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GNEVAR_TXT: }{A dataset with 6377 observations and the following
8 variables: manyID, Title, Beg, TreatyText, Source,
ieadbID, gnevarID, ecolexID.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $GNEVAR_TXT
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            6377|    207    |     3.25      |
## |     Title     |  character  |            6377|     0     |       0       |
## |      Beg      |    mdate    |            6377|    122    |     1.91      |
## |  TreatyText   |    list     |            6377|   3912    |     61.35     |
## |    Source     |    list     |            6377|   4749    |     74.47     |
## |    ieadbID    |  character  |            6377|   2657    |     41.67     |
## |   gnevarID    |  character  |            6377|   3587    |     56.25     |
## |   ecolexID    |  character  |            6377|   4167    |     65.34     |
## ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
