% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyenviron-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 1 datasets: AGR_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{AGR_TXT: }{A dataset with 6377 observations and the following
8 variables: manyID, Title, Beg, Text, Source, ieadbID, gnevarID, ecolexID.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\preformatted{## $AGR_TXT
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             6377  
## Number of columns          8     
## _______________________          
## Column type frequency:           
##   character                6     
##   list                     2     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ─────────────────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID              207         0.968     6    34     0     5794          0
## 2 Title                 0         1        15   550     0     5649          0
## 3 Beg                 122         0.981     4    10     0     4148          0
## 4 ieadbID            2657         0.583     1     4     0     3666          0
## 5 gnevarID           3587         0.438     9     9     0     2733          0
## 6 ecolexID           4167         0.347    10    10     0     2174          0
## 
## ── Variable type: list ──────────────────────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_unique min_length max_length
## 1 Text               3911         0.387     2123          0        287
## 2 Source             4749         0.255     1357          1          1
}
}
\keyword{datasets}
